#  storage/ api-docs/ api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Api Project",
    "description": "Demo my Project ",
    "contact": {
      "email": "admin@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/product": {
      "get": {
        "tags": [
          "Product"
        ],
        "operationId": "5cc57bec35d0f2045983cc8dcd00eccd",
        "responses": {
          "200": {
            "description": "Get All Products."
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "operationId": "1d0b0d8724685b7b8f9d87f403bfc4df",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "category_id",
                  "name",
                  "price",
                  "description",
                  "images[]"
                ],
                "properties": {
                  "category_id": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string"
                  },
                  "images[]": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Add Product."
          }
        }
      }
    },
    "/api/product/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "operationId": "b3bd2150e823630e583f0f6bb6446c81",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get a Product."
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "operationId": "118d6027a434cda88b95d382d657bd05",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Delete a Product."
          }
        }
      }
    },
    "/api/product/edit/{id}": {
      "post": {
        "tags": [
          "Product"
        ],
        "operationId": "6846909ff7036fa168f6ff283230a18c",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "category_id",
                  "name",
                  "price",
                  "description",
                  "images[]"
                ],
                "properties": {
                  "category_id": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string"
                  },
                  "images[]": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update a Product."
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "bearerFormat": "JWT",
        "scheme": "bearer"
      }
    }
  }
}